Changes as per these recommendations
XendCash heading is place sell order

On the Xend page, if you click the "How much can I send" button, it brings out a figure without proper description and coin symbol e.g. "You can send 0.16tBtc".

On the my order page, the page title is "Your Order" instead of "My Order" as it is stated on the menu button

The chat description does not call the seller to action. It only notifies the seller of a buyer's interest.

Can we add, please log in to your xendbit app to continue trade

Signed-off-by: Akintayo A. Olusegun <akintayo.segun@gmail.com>